LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audionode-interface/audionode-connect-return-value-expected.txt


FAIL connect should return the node connected to. assert_equals: expected (object) object "[object GainNode]" but got (undefined) undefined
